fixed up logging for database migrations
