+++ src/sophrosyne/commands/__init__.py
@@ -2,86 +2,17 @@
 
 #
 # Do NOT import any modules from sophrosyne before making sure you've read the
-# docstring of the _necessary_evil function.
+# docstring of the necessary_evil function.
 #
 
-import asyncio
 from contextlib import asynccontextmanager
-from functools import wraps
 
 import click
 from fastapi import FastAPI
 
-
-def _necessary_evil(path: str):
-    """Run some initial setup.
-
-    This code, as the name implies, is a necessary evil to make up for a
-    missing feature, or perhaps my own personal shortcomings, of Pydantic. It
-    does not seem possible to dynamically specify external configuration files
-    via `model_config` in Pydantic, forcing us to have the value of the
-    `yaml_file` argument be a variable that is set at module import time. This
-    unfortunately creates the side effect that the location of the yaml file
-    must be known the config module is imported.
-
-    The way this is handled is to have this function take care of setting the
-    necessary environment variables to configure the config module before
-    importing it.
-
-    It is imperative that this function is run before any other modules from
-    sophrosyne is imported. This is because many other modules import the
-    config module, and if that happens before this function is run, everything
-    breaks.
-
-    Additionally, because this function is run early and by pretty much all
-    commands, it is also used to centralize other things such as initialization
-    of logging.
-    """
-    import os
-
-    if "SOPH__CONFIG_YAML_FILE" not in os.environ:
-        os.environ["SOPH__CONFIG_YAML_FILE"] = path
-    import sophrosyne.core.config  # noqa: F401
-    from sophrosyne.core.config import get_settings
-    from sophrosyne.core.logging import initialize_logging
-
-    initialize_logging(
-        log_level=get_settings().logging.level_as_int,
-        format=get_settings().logging.format,
-        event_field=get_settings().logging.event_field,
-    )
-
-
-def async_cmd(func):
-    """Decorator to run an async function as a synchronous command.
-
-    This decorator allows you to use async functions as synchronous commands in Click.
-    It uses the `asyncio.run()` function to run the async function in a synchronous manner.
-
-    Args:
-        func (Callable): The async function to be decorated.
-
-    Returns:
-        Callable: The decorated function.
-
-    Example:
-        @async_cmd
-        async def my_async_command():
-            # async code here
-
-        if __name__ == "__main__":
-            my_async_command()
-
-    Reference:
-        This decorator is based on the solution provided in the following StackOverflow post:
-        https://stackoverflow.com/questions/67558717/how-can-i-test-async-click-commands-from-an-async-pytest-function
-    """
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        return asyncio.run(func(*args, **kwargs))
-
-    return wrapper
+from sophrosyne.commands import database
+from sophrosyne.commands.internal import necessary_evil
+from sophrosyne.core import async_cmd
 
 
 @asynccontextmanager
@@ -125,7 +56,7 @@ def version():
 )
 def config(config, pretty):
     """Print the configuration to stdout as JSON."""
-    _necessary_evil(config)
+    necessary_evil(config)
 
     from sophrosyne.core.config import get_settings
 
@@ -133,15 +64,15 @@ def config(config, pretty):
     if pretty:
         indent = 2
 
-    print(get_settings().model_dump_json(indent=indent))
+    click.echo(get_settings().model_dump_json(indent=indent))
 
 
 @click.command()
 @click.option("--config", default="config.yaml", help="path to configuration file.")
 @async_cmd
 async def healthcheck(config):
-    """Check the health of the SOPH API service."""
-    _necessary_evil(config)
+    """Check the health of the sophrosyne API service."""
+    necessary_evil(config)
 
     import sys
 
@@ -151,9 +82,6 @@ async def healthcheck(config):
     from sophrosyne.core.database import (
         engine,
     )
-    from sophrosyne.core.logging import get_logger
-
-    log = get_logger()
 
     # Disable warnings for insecure requests
     requests.packages.urllib3.disable_warnings()
@@ -168,7 +96,7 @@ async def healthcheck(config):
             verify=verify,
         )
         if resp.status_code != 200 and resp.text != '"pong"':
-            log.error("API returned abnormal response.")
+            click.echo("API returned abnormal response.")
             return sys.exit(1)
     except requests.exceptions.ConnectionError as e:
         # This is not really a nice way of doing this, is there not a better way?
@@ -177,9 +105,9 @@ async def healthcheck(config):
             reason = reason.removeprefix("certificate verify failed: ")
             reason = reason[: reason.rfind(" (")]
             reason = reason.strip()
-            log.error(f"SSL/TLS verification failure: {reason}")
+            click.echo(f"SSL/TLS verification failure: {reason}")
         else:
-            log.error("API is not responding.")
+            click.echo("API is not responding.")
         return sys.exit(1)
 
     from sqlalchemy.ext.asyncio import async_sessionmaker
@@ -195,17 +123,17 @@ async def healthcheck(config):
     async with db_session() as session:
         hc = await do_authenticated_healthcheck(db_session=session)
         if hc.status == "pass":
-            log.info("The server is healthy.")
+            click.echo("The server is healthy.")
         else:
-            log.error("The server is not healthy.")
+            click.echo("The server is not healthy.")
             return sys.exit(1)
 
 
 @click.command()
 @click.option("--config", default="config.yaml", help="path to configuration file.")
 def run(config):
-    """Run the SOPH API service."""
-    _necessary_evil(config)
+    """Run the sophrosyne API service."""
+    necessary_evil(config)
 
     import sys
 
@@ -264,3 +192,19 @@ def run(config):
         # uvicorn, but it works as expected.
         ssl_keyfile_password=tls.private_key_password,  # type: ignore
     )
+
+
+@click.group()
+def root():
+    """Sophrosyne - A content moderation API."""
+    pass
+
+
+def setup_and_run_commands():
+    """Setup the CLI commands and execute the root command."""
+    root.add_command(version)
+    root.add_command(run)
+    root.add_command(healthcheck)
+    root.add_command(config)
+    root.add_command(database.cmd)
+    root()
